start wrok on undo redo
